FaceDetection: new implementation (integrated!)
When running the program in Run.java and clicking 'Proceed' in the welcome window it will start the face detection in the background ! It does not open a video stream but saves an image of the detected face! Can probably make the stream work too though. Currently only face detection, no identification yet.